Changed user interface
