+++ app/src/main/AndroidManifest.xml
@@ -13,15 +13,12 @@
         android:screenOrientation="portrait"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.login.activity.RegAndSingActivity">
+        <activity android:name=".utils.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity android:name=".utils.SplashActivity">
-
         </activity>
         <activity android:name=".MainActivity" />
