--- app/src/main/java/com/example/movieapp/ui/login/registration/RegistrationViewModel.kt
@@ -1,7 +1,0 @@
-package com.example.movieapp.ui.login.registration
-
-import androidx.lifecycle.ViewModel
-
-class RegistrationViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
