--- app/src/main/java/com/example/movieapp/ui/login/sing_in/SingInViewModel.kt
@@ -1,7 +1,0 @@
-package com.example.movieapp.ui.login.sing_in
-
-import androidx.lifecycle.ViewModel
-
-class SingInViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
