+++ app/src/main/java/com/example/movieapp/utils/SplashActivity.kt
@@ -7,7 +7,7 @@ import android.os.Handler
 import androidx.appcompat.app.AppCompatActivity
 import com.example.movieapp.MainActivity
 import com.example.movieapp.R
-import com.example.movieapp.ui.login.activity.RegAndSingActivity
+
 import kotlinx.android.synthetic.main.activity_splash.*
 
 class SplashActivity : AppCompatActivity() {
@@ -24,7 +24,7 @@ class SplashActivity : AppCompatActivity() {
         }
 
         Handler().postDelayed({
-            startActivity(Intent(this, RegAndSingActivity::class.java))
+            startActivity(Intent(this, MainActivity::class.java))
             finish()
         }, 1500L)
     }
