--- app/src/main/res/navigation/reg_and_sign.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/reg_and_sign"
-    app:startDestination="@id/registrationFragment">
-
-    <fragment
-        android:id="@+id/registrationFragment"
-        android:name="com.example.movieapp.ui.login.registration.RegistrationFragment"
-        android:label="RegistrationFragment"
-        tools:layout="@layout/registration_fragment">
-        <action
-            android:id="@+id/action_registrationFragment_to_singInFragment"
-            app:destination="@id/singInFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/singInFragment"
-        android:name="com.example.movieapp.ui.login.sing_in.SingInFragment"
-        android:label="SingInFragment"
-        tools:layout="@layout/sing_in_fragment">
-        <action
-            android:id="@+id/action_singInFragment_to_registrationFragment"
-            app:destination="@id/registrationFragment" />
-    </fragment>
-</navigation>
